requirements.txt
Django~=4.2.4
pip
psycopg2==2.9.7
djangorestframework~=3.14
python-dotenv~=1.0.0
drf-yasg~=1.21.7
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.26.4
Pillow==10
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
pep8-naming==0.13.2
psycopg2
pytz
sqlparse
python-dotenv~=1.0.0
Pillow~=10.0.0
django-debug-toolbar==4.2.0
